WIP: filled out a bit more of the Driver API
need to rework the decoder to have multiple output queues before I can
make meaningful progress on any of the calls that involve upstream
traffic. Probably going to get rid of or alter heavily the Xcoder base
class in the process.

BDState needs to be filled out. Some calls are impossible without
tracking the present state, and all programming calls need to be modified
to capture programmed state.