Got around gcc defect by changing unordered_map -> vector
Older versions of gcc can't use enums as unordered_map keys.
In the situation I was using the unordered_map I probably should have
been using a vector anyway, so I switched to one.

Also made a bunch of small changes to squelch some compiler warnings.
Obnoxiously, assert() (perhaps only in release mode... which is kind of
appropriate since it's a nop() in that case) can cause a bunch of
warnings because the compiler isn't smart enough to detect that
assert(false) is the end of execution. Also, variables used inside
assert()s don't count as being used in this case.