Started working on BDModel
not integrated into the build yet!

Trying to parse the memory words turned out to be a nightmare (daymare?).
It was probably about as much work as the driver calls themselves.

I'm not totally happy with it, it's pretty damn complicated. Maybe
that's just the way it has to be, though.

BDModelUtils might be useful to other testing code.
Alternatively, I probably could have made these static functions of BDModel.

The only thing left to do is iterate through the VFVVs for each leaf and
update the BDState object.