Big simplification of BDModel
Dodges all the VectorFieldVValues/Collapse/Deencapsulate stuff.
The Parse/Process steps in BDModel have been unified by eliminating
the old Parse logic from BDModelUtil and using Driver::Unpack in a much
more direct way.

That stuff might have actually worked (there were some fundamental
bugs in Pack32->64 that were killing me), but it was a lot of added
complexity that was more trouble to verify than it was worth. It was a
pretty confusing set of abstractions to begin with.

I will probably purge the BDModelUtil stuff which is unused in a future
commit.