Streamlined implementation of Pack/UnpackWord(s)
logic was repeated for each. The FVV versions of these calls aren't
even that important: they aren't used for performance anywwhere.

FVV is a possible candidate for purging. I don't think it's ever used
in a place where performance is needed (which was the original point of
creating it). The conversion functions are kind of a pain.