Fix options for building with Clang
The Clang compiler was not being passed build options because the
if condition was only checking whether a GNU CXX compiler was
being used. To fix this, the CMakeLists file checks whether
the Clang compiler is being used when building on Mac and if so,
the CMAKE_CXX_COMPILER_ID will be equal to AppleClang.

In addition, I removed some flags which are ignored by the Clang compiler
and moved the Jenkinsfile back so that it can be built on Jenkins.