Use library names common to Mac and Linux platforms (#20)
* Use library names common to Mac and Linux platforms

The PREFIX property for the bddriver library target was overridden and
a link from a generic file to the built library was created to
provide a generic name for the libraries that was platform independent.

This was the incorrect way of using CMake for platform independent builds.
The correct approach was to NOT override the suffix and prefix, add
link directories (using CMake's add_link_directory command) and
reference the library by the given name.

The compiler will figure out the appropriate prefix and suffix given
just the library name. For example, GCC on Linux will figure out that
the linked library Driver is mapped to the file libDriver.so and
GCC on Mac will figure out that the linked library Driver is mapped to
the file libDriver.dyld.

* Fix options for building with Clang

The Clang compiler was not being passed build options because the
if condition was only checking whether a GNU CXX compiler was
being used. To fix this, the CMakeLists file checks whether
the Clang compiler is being used when building on Mac and if so,
the CMAKE_CXX_COMPILER_ID will be equal to AppleClang.

In addition, I removed some flags which are ignored by the Clang compiler
and moved the Jenkinsfile back so that it can be built on Jenkins.