WIP: refactor mostly done, code compiles, but having linking problems
My design with a struct for each word probably won't work. Since the
::field_widths and ::field_hard_values members aren't actually integral
type, they have to be declared in any .cpp that uses them.

Even though they only show up in BDPars.h::BDWord, it seems like the
symbol has to be declared in other translation units that are using it.

I may do this for now, but it's an undesirable requirement on the user.
Kinda crappy to have to declare MyWordName::field_widths/field_hard_values,
which you wouldn't even have been aware of, otherwise.

Will probably switch to an approach that uses pure enums.

(structured) compile-time constants are hard...