Refactor Driver to Use New BDWord Class (#31)
* Made driver diagram comment prettier

* Working version of new BDWord concept

* WIP: refactor mostly done, code compiles, but having linking problems

My design with a struct for each word probably won't work. Since the
::field_widths and ::field_hard_values members aren't actually integral
type, they have to be declared in any .cpp that uses them.

Even though they only show up in BDPars.h::BDWord, it seems like the
symbol has to be declared in other translation units that are using it.

I may do this for now, but it's an undesirable requirement on the user.
Kinda crappy to have to declare MyWordName::field_widths/field_hard_values,
which you wouldn't even have been aware of, otherwise.

Will probably switch to an approach that uses pure enums.

(structured) compile-time constants are hard...

* WIP: everything almost back together, only mem dump tests failing

still have to address the issue with the redeclaration of
SomeWord::field_widths/field_hard_values

* Refactor working (not done yet). Tests pass.

Biggest problem was fixing a typo in BDState::SetMem().

Also had a bug in how SomeWord::len was being used. Added ::width.
(This will be overhauled again when I probably move to enums).