Use correct connection sizes when associating Connectable objects with Connection objects
When connecting a Pool to a Bucket with a Connection object, we found out
that the connection was using the Pools dimensions and the Buckets dimensions
as the required dimensions of Connection objects weight matrix.

This was incorrect.

The Pools number of neurons indicates the input dimenions of the Connection
object and the weight matrix's number of columns should equal the
Pools number of neurons.

If a Pool is the input to a Connection, the number of neurons is the Connections
input size. If the input to a Connection is an Input or Bucket, the objects
number of dimensions is the Connections input size. To implement this requirement,
the Connection interface was given a GetConnectionSize method and the Pool
class returns the number of neurons whereas the other Connectable objects
return the number of dimensions.