Minimum timeout for OK calls, fixes deadlock
after fixing the FPGA bug where we were throwing away packets,
we caused a deadlock because the FPGA wasn't a perfect sink anymore
(it's not supposed to be).

The single-duplex nature of the OK causes problems when your input
words can elicit output words. With the default timeouts,
WriteToBlockPipeIn is effectively a blocking call.

Changed CommOK to only transmit one block downstream per write call as well:
this is possibly a throughput penalty and should be fixed