Fixes SetTagTrafficState, adds GetOutputQueueCounts
SetTagTrafficState was modifying the neuron array injection,
which is supposed to only be controlled by SetSpikeTrafficState

also modifies the ResetBD, which isn't perfect. It's hard to ensure
that the states are held for an adequate amount of time,
since there's no good way to guarantee a delay in our timing scheme.