Deepened OKIfc FIFOs. Optimized upstream behavior.
FPGA now uses a BTPipe and sends words like this: ddddXXXXXX
where d is data and X is nop. Previously, this was possible:
ddddXXXdXXXdX. This means that the Decoder can now ignore words in a
block after the first nop. This is important because the faster we go,
the more the decoder becomes a bottleneck.