Updated BDPars to properly scale ADC currents
I also added comments to document the approximate current values for any
bias, after scaling from 1pa/1nA down.

This commit also includes the latest version of the test_adc.py script
which toggles the various modes of the ADCs.  The next step of this
script will be to test varying the input current to the ADC.