increment sleeps in test recv spikes and tags
TestRecvSpikes and TestRecvTags test that the spike and tags
sent from the driver to the software model of braindrop match
the subsequent spikes and tags received from the model.
To wait for the model, the tests sleep for a time before checking
the received and sent data. The build fails if the timing assumption
is violated.