Added more descriptive labeling to parse_power_data.py
I also updated measure_power.py to have a for loop inside of the
DecodeEncode run that allows me to vary temperature and read out firing
rates, tag rates, and fifo overflows.  This code is commented so by
default, so that the next person who checks out this code doesn't get
unexpected behavior, but they can now uncomment it if they want to
measure repeated values over a longer period of time.