test_clobbering.py is not pytest compatible with this version
I commented out the pytest compatibility for the time being to keep the
code pushing through any errors for board debugging purposes.  I was
tired of stashing and unstashing, so I'm commiting to this repo for now,
so that the default for testing the bad_grt bits is this version of the
code.