Made test_sticky_bits pytest compatible
The scripts assumes that even entries of the memory array have all 0
values and odd entries have all 1 values.  It checks that they output
the correct bit strings, and asserts if there is ever any discrepancy.

It will report the memory entry, the bit index, and the value of the
incorrect bit as stderr output.