test_sequenced_sticky_bits before trying to interleave tags
This is a working version of the test_sequenced_sticky_bits test used to
characterize the clobbering vs sticky bits problem.  It was able to suss
out the fact that these two problem occur in different scenarios.  The
next step is to check if the sticky bits problem occurs when we try to
recreate a tag pattern similar to how a dump memory output occurs.