Power measurement scripts updated to measure AERTX alone and AERRX
Note the AERRX measurements don't work perfectly in this script, as they
take too long to send data to the driver.  As such, this is the state of
the system before I create a new branch to try and resolve this slow
data processing, which is believed to be caused by copying a large array
from python to C each time.