calibration/syn_tau.py now more reliable about NaN
before fitting the combined responses of the neurons under each synapse to
estimate tau, the script needs to make sure that the neuron responses actually
changed in the first place (they might not if none of the nearby neurons had
linear responses over the 0-FMAX range, for instance). Otherwise, it could try
to fit a decay to meaningless (noisy constant value) data.