verification for syn_tau.py calibration
The indexing/collapsing of the data in syn_tau.py is pretty tricky.
validate_syn_tau.py does the same algorithm as syn_tau.py, but only measures
one synapse at a time (instead of 1/4 of the entire chip). The script checks the values it obtains against those in the database.